% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v.pwr.p.test.R
\name{v.pwr.p.test}
\alias{v.pwr.p.test}
\title{Vectorizes power calculations for proportion tests (one sample).}
\usage{
v.pwr.p.test(h = NULL, n = NULL, sig.level = 0.05, power = NULL,
  alternative = c("two.sided", "less", "greater"), deff = 1)
}
\arguments{
\item{h}{Effect size, could input a single numerical value or a vector of
values, applies to all parameters.}

\item{n}{Number of observations.}

\item{sig.level}{Significant level.}

\item{power}{Power of test.}

\item{alternative}{A character string specifying the alternative hypothesis.
Must be one of "two.sided" (default), "greater" or "less".}

\item{deff}{Design effect.}
}
\value{
A list of sublists contain values(s) of the arguments augmented with
  'method' elements. The argument(s) are assigned/computed
  multiple values will be returned in a sequence of expanding combination,
  while the argument(s) inputted single value will be returned only one
  single value.
}
\description{
\code {v.pwr.p.test} computes power of test or determine paramaters to obtain
target power. It allows up to 2 vectorize arguments.
}
\section{Warning}{
 The function only takes up to 2 vectorize arguments.
}

