% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v.pwr.t.test.R
\name{v.pwr.t.test}
\alias{v.pwr.t.test}
\title{Vectorizes power calculations for t-test of means.}
\usage{
v.pwr.t.test(n = NULL, d = NULL, sig.level = 0.05, power = NULL,
  type = c("two.sample", "one.sample", "paired"),
  alternative = c("two.sided", "less", "greater"), deff = 1)
}
\arguments{
\item{n}{Number of observations.}

\item{d}{Effect size, - difference between the means divided by the pooled
standard deviation. Could input a single numerical value or a vector of
values, applies to all parameters.}

\item{sig.level}{Significant level.}

\item{power}{Power of test.}

\item{type}{Type of t test: one- two- or -paired-samples. Two samples is
default.}

\item{alternative}{A character string specifying the alternative hypothesis.
Must be one of "two.sided" (default), "greater" or "less".}

\item{deff}{Design effect.}
}
\value{
A list of sublists contain values(s) of the arguments augmented with
  'method' and 'NOTE' elements. The argument(s) are assigned/computed
  multiple values will be returned in a sequence of expanding combination,
  while the argument(s) inputted single value will be returned only one
  single value.
}
\description{
\code {v.pwr.t.test} computes power of test or determine paramaters to obtain
target power. It allows up to 2 vectorize arguments.
}
\section{Warning}{
 The function only takes up to 2 vectorize arguments.
}

