% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_pwr.R
\name{vector_pwr}
\alias{vector_pwr}
\title{Backend vectorizer for functions in the pwr package}
\usage{
vector_pwr(..., deff)
}
\arguments{
\item{...}{Any parameters passed to a function in the \\code{pwr} package.}

\item{deff}{The design effect}
}
\value{
An \\code{rbind} of the result of applying the appropriate function
    from the \\code{pwr} package to each combination of the input parameters.
}
\description{
Backend vectorizer for functions in the pwr package
}
\section{Warning}{
 It allows up to 2 vectorized arguments
}

